Fix race condition between obstacles and transforms
tf is published by the localization stack, obstacles are published by
the perception stack. There is a race condition where the obstacle is
published first shortly before tf is published. In which case obstacles
cannot be transformed into world frame straight away.

In this patch a timer is used to periodically check if a valid transform
is available before performing the actual transform.

Change-Id: I237f38f53fd4a654f04ef80227f062b0966e29b2