[parking_planner_node] Fixed unions being miscomputed.
As described in https://gitlab.com/autowarefoundation/autoware.auto/AutowareAuto/-/issues/636#note_417806285,
CGAL seems to have a property that "union(something,nothing)"
results in "nothing", instead of "something" as I would have
assumed. This commit works around this, but as a result, if
the primitives don't come out correctly yet, this will now mean
that the planner can fail to come up with a feasible trajectory
(rightfully so).