[parking_planner_node] Found and fixed another culprit for wonky planning
Found out that I forgot to change the rotation of points from clockwise
to counter-clockwise in a critical point, leading to the obstacle
halfplane constraints being "inside out". Due to the mathematical
formulation of the problem, this was not apparent in the behavior so
far, but it is correct now.